Изменил миграцию

+++ migrations/20250112193902_init_schema.sql
@@ -3,8 +3,8 @@
 CREATE TABLE metrics (
     id TEXT PRIMARY KEY,
     metric_type VARCHAR(10) NOT NULL CHECK (metric_type IN ('gauge', 'counter')),
-    value FLOAT,
-    delta INT
+    value DOUBLE PRECISION,
+    delta BIGINT
 );
 -- +goose StatementEnd
